+++ src/lib.rs
@@ -10,7 +10,7 @@
 //! ## Usage
 //!
 //! Run
-//! 
+//!
 //! ```ignore
 //! $ cargo add groan_rs
 //! ```